% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{craft2MEMUNITS}
\alias{craft2MEMUNITS}
\title{convert Craft to logical memory iia
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111}
\usage{
craft2MEMUNITS(craft)
}
\arguments{
\item{mlnttots}{multilingual naming scores as vector}
}
\value{
boston naming test scores as vector
}
\description{
convert Craft to logical memory iia
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111
}
\examples{
craft2MEMUNITS(c(23,124,53))
}
