% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{digbacct2DIGIB}
\alias{digbacct2DIGIB}
\title{convert digbacct to DIGIB
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111}
\usage{
digbacct2DIGIB(digbacct)
}
\arguments{
\item{digbacct}{0-14    High score = good}
}
\value{
DIGIB #WMS-R DIGIT SPAN BACKWARD (total number of trials)	Num 	0-12	High score = good
}
\description{
convert digbacct to DIGIB
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111
}
\examples{
digbacct2DIGIB(c(0,7,53))
}
