% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{digforct2DIGIF}
\alias{digforct2DIGIF}
\title{convert digforct to DIGIF
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111}
\usage{
digforct2DIGIF(digforct)
}
\arguments{
\item{digforct}{WMS Digit Span Forward    Num     0-14    High score = good}
}
\value{
DIGIF #WMS-R DIGIT SPAN FORWARD (total number of trials)	Num 	0-12	High score = good
}
\description{
convert digforct to DIGIF
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111
}
\examples{
digforct2DIGIF(c(0,7,53))
}
