% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_ADCSPACC}
\alias{get_ADCSPACC}
\title{Takes in dataframe of normed scores
Returns ADCS PACC
index_srtfree,index_craftdel,index_digsym, index_mmse}
\usage{
get_ADCSPACC(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  srtfree_col_name = "z_srtfree",
  craftdel_col_name = "z_craftdre",
  mmse_col_name = "z_MMSE_eq",
  digsym_col_name = "z_digsym"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{srtfree_col_name}{character string of srtfree column name}

\item{digsym_col_name}{character string of digit symbol column name}
}
\value{
dataframe containing participant id, visit date, the global composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns ADCS PACC
index_srtfree,index_craftdel,index_digsym, index_mmse
}
