% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_AttentionComposite}
\alias{get_AttentionComposite}
\title{Takes in dataframe of normed scores
Returns Attention Processing Composite
index_tma,index_tmb,index_digsym, index_switch}
\usage{
get_AttentionComposite(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  tma_col_name = "z_TrailA_eq",
  tmb_col_name = "z_TrailB_eq",
  digsym_col_name = "z_digsym",
  switch_col_name = "z_switchmixed"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{tma_col_name}{character string of trails a column name}

\item{tmb_col_name}{character string of trails b column name}

\item{digsym_col_name}{character string of digit symbol column name}

\item{switch_col_name}{character string of switch column name}
}
\value{
dataframe containing participant id, visit date, the attention processing composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns Attention Processing Composite
index_tma,index_tmb,index_digsym, index_switch
}
