% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataNorming.R
\name{get_baselineCohort}
\alias{get_baselineCohort}
\title{Takes in dataframe of cognitive data
Returns dataframe of baseline scans for cognitively normal participants for norming}
\usage{
get_baselineCohort(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  cdr_score_name = "cdr"
)
}
\arguments{
\item{df}{dataframe containing participant id, visit date, cdr score, and relavant cognitive tests}
}
\value{
df corresponding dataframe with only baseline visits
}
\description{
Takes in dataframe of cognitive data
Returns dataframe of baseline scans for cognitively normal participants for norming
}
\examples{
df <- data.frame("ID" = c(1, 1, 2, 3), "cdr" = c(0, 0, 0,5, 0), 
                  "TESTDATE" = c("1989-04-11", "1991-03-26", "2011-05-31", "2021-06-23"),
                  "ANIMALS" = c(28, 27, 20, 11))
get_baselineCohort(df, "ID", "TESTDATE", "cdr")
}
