% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_EpisodicMemoryComposite}
\alias{get_EpisodicMemoryComposite}
\title{Takes in dataframe of normed scores
Returns Episodic Memory Composite}
\usage{
get_EpisodicMemoryComposite(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  craftdel_col_name = "z_craftdre",
  asscmem_col_name = "z_asscmem",
  srtfree_col_name = "z_srtfree"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{craftdel_col_name}{character string of craftdel column name}

\item{asscmem_col_name}{character string of associated memory column name}

\item{srtfree_col_name}{character string of srtfree column name}
}
\value{
dataframe containing participant id, visit date, the episodic memory composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns Episodic Memory Composite
}
