% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_GlobalComposite}
\alias{get_GlobalComposite}
\title{Takes in dataframe of normed scores
Returns Global Composite}
\usage{
get_GlobalComposite(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  craftdel_col_name = "z_craftdre",
  asscmem_col_name = "z_asscmem",
  srtfree_col_name = "z_srtfree",
  animals_col_name = "z_animals",
  mlnt_col_name = "z_mlnt_eq",
  veg_col_name = "z_veg",
  tma_col_name = "z_TrailA_eq",
  tmb_col_name = "z_TrailB_eq",
  digsym_col_name = "z_digsym",
  switch_col_name = "z_switchmixed",
  digb_col_name = "z_DIGIB_eq",
  digf_col_name = "z_DIGIF_eq",
  lettnum_col_name = "z_lettnum"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{craftdel_col_name}{character string of craftdel column name}

\item{asscmem_col_name}{character string of associated memory column name (WMS Associate Learning (summary score)    Num     0-21    High score = good)}

\item{srtfree_col_name}{character string of srtfree column name}

\item{animals_col_name}{character string of animals column name  (CATEGORY FLUENCY - ANIMALS    Num     0 and above    High score = good)}

\item{mlnt_col_name}{character string of multilingual naming test column name  (Multilingual Naming Test (MINT) - Total Score    Num     0-32    High Score = good)}

\item{veg_col_name}{character string of veg column name (CATEGORY FLUENCY - VEGETABLES    Num     0 and above    High score = good)}

\item{tma_col_name}{character string of trails a column name (High score = bad)}

\item{tmb_col_name}{character string of trails b column name (High score = bad)}

\item{digsym_col_name}{character string of digit symbol column name (WAIS-R Digit Symbol    Num     0-93    High score = good)}

\item{switch_col_name}{character string of switch column name}

\item{digb_col_name}{character string of digb column name (WMS-R DIGIT SPAN BACKWARD (total number of trials)    Num     0-12    High score = good)}

\item{digf_col_name}{character string of digf column name #WMS-R DIGIT SPAN FORWARD (total number of trials)    Num     0-12    High score = good}

\item{lettnum_col_name}{character string of lettnum column name #WMS-III letter-Number sequencing    Num     0-21    High score = good}
}
\value{
dataframe containing participant id, visit date, the global composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns Global Composite
}
