% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_SemanticMemoryComposite}
\alias{get_SemanticMemoryComposite}
\title{Takes in dataframe of normed scores
Returns Semantic Memory
index_animals,index_mint,index_veg}
\usage{
get_SemanticMemoryComposite(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  animals_col_name = "z_animals",
  mlnt_col_name = "z_mlnt_eq",
  veg_col_name = "z_veg"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{animals_col_name}{character string of animals column name}

\item{mlnt_col_name}{character string of multilingual naming test column name}

\item{veg_col_name}{character string of veg column name}
}
\value{
dataframe containing participant id, visit date, the semantic memory composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns Semantic Memory
index_animals,index_mint,index_veg
}
