% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{get_TRAILSA}
\alias{get_TRAILSA}
\title{combine two vectors that contain Trails A data to get a single trails A vector}
\usage{
get_TRAILSA(tma, TRAILA)
}
\arguments{
\item{tma}{trails a (number of seconds)}

\item{TRAILA}{TRAILMAKING A (number of seconds)    Num     0-150}
}
\value{
TRAILA A combination of the two columns
}
\description{
combine two vectors that contain Trails A data to get a single trails A vector
}
\examples{
get_TRAILSA(c(NA,7,53), c(300, NA, NA))
}
