% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{get_TRAILSB}
\alias{get_TRAILSB}
\title{combine two vectors that contain Trails B data to get a single trails B vector}
\usage{
get_TRAILSB(tmb, TRAILB)
}
\arguments{
\item{tmb}{trails b (number of seconds)}

\item{TRAILB}{TRAILMAKING B (number of seconds)    Num     0-300}
}
\value{
TRAILB A combination of the two columns
}
\description{
combine two vectors that contain Trails B data to get a single trails B vector
}
\examples{
get_TRAILSB(c(NA,507,53), c(300, NA, NA))
}
