% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreComposites.R
\name{get_WorkingMemoryComposite}
\alias{get_WorkingMemoryComposite}
\title{Takes in dataframe of normed scores
Returns Working Memory Composite
index_digb,index_digf,index_lettnum}
\usage{
get_WorkingMemoryComposite(
  df,
  id_col_name = "ID",
  visit_date_col_name = "TESTDATE",
  digb_col_name = "z_DIGIB_eq",
  digf_col_name = "z_DIGIF_eq",
  lettnum_col_name = "z_lettnum"
)
}
\arguments{
\item{df}{dataframe containing normed scores}

\item{id_col_name}{character string of ID column name}

\item{visit_date_col_name}{character string of visit date column name}

\item{digb_col_name}{character string of digb column name}

\item{digf_col_name}{character string of digf column name}

\item{lettnum_col_name}{character string of lettnum column name}
}
\value{
dataframe containing participant id, visit date, the global composite score and the number of nas in the visit that went into the score. It is up to user discretion in terms of how many tests to require
}
\description{
Takes in dataframe of normed scores
Returns Working Memory Composite
index_digb,index_digf,index_lettnum
}
