% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{mlnt2BNT}
\alias{mlnt2BNT}
\title{convert multilingual naming test to boston naming test equivalent
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111}
\usage{
mlnt2BNT(mlnttots)
}
\arguments{
\item{mlnttots}{multilingual naming scores as vector}
}
\value{
boston naming test scores as vector
}
\description{
convert multilingual naming test to boston naming test equivalent
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111
}
\examples{
minttots2BNT(c(23,124,53))
}
