% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testEquivalents.R
\name{MoCA2MMSE}
\alias{MoCA2MMSE}
\title{convert Moca scores to mmse
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111}
\usage{
MoCA2MMSE(MoCA)
}
\arguments{
\item{MoCA}{MoCa scores as vector}
}
\value{
equivalent MMSE scores as vector
}
\description{
convert Moca scores to mmse
Crosswalks from Monsell et al, 2016 DOI: 10.1097/WAD.0000000000000111
}
\examples{
MoCA2MMSE(c(23,124,53))
}
